backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==24.3
PyMySQL==1.1.1
cryptography==42.0.8
python-dotenv==1.0.0
django-cors-headers==4.2.0
drf-yasg==1.21.7
Pillow==10.1.0
